problem: MyEq hessian function
Changed NEQ and NINEQ variables in optizellescopf_solver.m.

NEQ = 2*nb + 2*nl now
and
NINEQ = 2*length(x) (before updating with slack variables)

Now, code in hessian of MyEq is not working. Pretty sure it's because of lam.ineqnolin and lam.eqnolin but not sure how to fix it at the moment.